Udate Dockerfile: dependencies before source code
Due to the nature of docker build caching, if pip is run after the source code is added to the container any change to the source code directories (blog, and test) will invalidate the cache entry for the pip RUN statement, thus installing all dependencies each time a build with source changes is run. This is slow, only run pip when project dependencies change.